refactor code based on reviews
