+++ modflowapi/__init__.py
@@ -1,5 +1,5 @@
 # imports
 from .version import __version__
 from modflowapi.modflowapi import ModflowApi
-from . import interface
-from .interface.runner import run_model, Callbacks
+from . import extensions
+from .extensions.runner import run_simulation, Callbacks
