+++ modflowapi/extensions/__init__.py
@@ -1,0 +1,2 @@
+from .apisimulation import ApiSimulation
+from .apimodel import ApiModel
