--- modflowapi/interface/__init__.py
@@ -1,2 +1,0 @@
-from .simulation import Simulation
-from .model import Model
